apps/backend/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
django-cors-headers
requests
djangorestframework-simplejwt
pyotp
qrcode
pillow
django-rest-passwordreset
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pygsheets
pandas
